openharmony-schedule - 42

+++ openharmony/arkui_ace_engine/frameworks/core/components_ng/pattern/list/list_position_map.cpp
@@ -54,7 +54,7 @@ ListPosMapUpdate ListPositionMap::CheckPosMapUpdateRule()
     return flag;
 }
 
-void ListPositionMap::UpdatePosMapStart(float delta, float& listCurrentPos, float space,
+void ListPositionMap::UpdatePosMapStart(float delta, double& listCurrentPos, float space,
     int32_t startIndex, float startPos, bool groupAtStart)
 {
     auto it = posMap_.find(startIndex);
