+++ openharmony/arkui_ace_engine/frameworks/core/components_ng/pattern/list/list_position_map.h
@@ -93,7 +93,7 @@ public:
 
     ListPosMapUpdate CheckPosMapUpdateRule();
 
-    void UpdatePosMapStart(float delta, float& listCurrentPos, float space,
+    void UpdatePosMapStart(float delta, double& listCurrentPos, float space,
         int32_t startIndex, float startPos, bool groupAtStart);
 
     void UpdatePosMapEnd(int32_t prevEndIndex, float space, bool groupAtEnd);
