+++ openharmony/arkui_ace_engine/test/unittest/core/pattern/web/web_pattern_focus_test_ng.cpp
@@ -265,7 +265,7 @@ public:
 
     void StopAnimate() override {}
 
-    float GetTotalOffset() const override
+    double GetTotalOffset() const override
     {
         return 0.0f;
     }
